requirements.txt
mysql-connector-python
alpha_vantage
pandas
Flask
